+++ src/shared/backend/replicon_channels.rs
@@ -98,7 +98,7 @@ impl RepliconChannels {
 /// are split into packet-size messages to allow applying them partially without waiting for all parts of the message.
 /// We guarantee that all mutations for a single entity arrive together, even if they are larger than the packet size.
 /// You can also ensure that mutations for specific entities arrive together by using
-/// [`ReplicateTogetherAppExt::replicate_together`](crate::shared::replication::related_entities::ReplicateTogetherAppExt::replicate_together).
+/// [`SyncRelatedAppExt::sync_related_entities`](crate::shared::replication::related_entities::SyncRelatedAppExt::sync_related_entities).
 ///
 /// Server events also have minimum required tick. For details, see the documentation on
 /// [`ServerEventAppExt::make_independent`](crate::shared::event::server_event::ServerEventAppExt::make_independent).